+++ Arnible.MathModeling/NumberMath.cs
@@ -2,8 +2,8 @@
 {
   public static class NumberMath
   {
-    public static Number Sin(Number a) => IndeterminateExpression.Sin(a);
+    public static Number Sin(Number a) => IndeterminateExpression.Sin((double)a);
 
-    public static Number Cos(Number a) => IndeterminateExpression.Cos(a);
+    public static Number Cos(Number a) => IndeterminateExpression.Cos((double)a);
   }
 }
